feat: primitive array refactor + feedback
