+++ packages/dm-core-plugins/src/form/templates/shared/FormExpandedViewWrapper.tsx
@@ -2,7 +2,7 @@ import React, { PropsWithChildren } from 'react'
 
 const FormExpandedViewWrapper = ({ children }: PropsWithChildren) => {
   return (
-    <div className='border-t p-2 border-[#6f6f6f] overflow-scroll'>
+    <div className='border-t p-2 border-[#6f6f6f] overflow-auto w-full'>
       {children}
     </div>
   )
