+++ packages/dm-core/src/components/index.tsx
@@ -6,6 +6,7 @@ export * from './JsonView'
 export * from './Loading'
 export * from './NewEntityButton'
 export * from './DeleteHardButton'
+export * from './AddRowButton'
 export * from './Pagination/Pagination'
 export * from './Pickers'
 export * from './Select'
